requirements.txt
gunicorn==21.2.0
Flask==3.1.0
Pillow==11.0.0
pdf2image==1.17.0
click==8.1.8
Jinja2==3.1.5
Werkzeug==3.1.3
